Fix a bug that breaks shooting while holding jump button.
This is an attempt to upstream a bugfix from Q3VR repo to Quake3Quest.

The problem here is that wrong field is compared with PM_SPECTATOR to
test if we're spectating now. Instead of comparing it with `pm_type`,
the original code compared it with `pm_flags`, which use different enums
(`PMF_...`). Due to lack of strongly-typed enums this was unnoticed.
It also unfortunate because `PM_SPECTATOR` gets value 3, which
corresponds to `(PMF_DUCKED & PMF_JUMP_HELD)` flags. So if we've jumped
(and have not let go of the jump button yet) and _ducked_ on land, this
will break aiming logic and projectiles will be fired in a wrong
direction.